Check for timezone changes and re-exec.
For now we're back to linux-only, due to use of AT_EMPTY_PATH &c., but
we can change our approach later to use the more portable stat(2) and
lstat(2) functions.

We handle the timezone change by re-execing ourselves, which should
reload and see any changes nicely. As a bonus, it'll pick up on new
executable changes, too.

Yeah, the re-exec is unnecessary for glibc, but other libc
implementations might not be so courteous. Plus what's it hurt, 2 extra
fstat(2)s per minute?